Move printrun/plater.py to printrun/stlplater.py
